Create level 1
this starts off as a first person player game. You start on a black block and are required to jump on to the diagonally opposite block on the platform in order to end the game. When you jump on a block it will either turn green or red. If it turns red then you have 2 seconds to jump on a different block before it breaks. Incase it turns green then the block is safe and won't break.