Create level 2
This level is similar to the level one except that there are added obstacles on the platform. These obstacles are huge entities that go from one end of the platform to another and back at a random pattern. They will sweep you off the platforms and its necessary to avoid them in order for you to continue playing.