Create level 3
This is the last of the levels and has 2 major additions to the previous level. First the blocks are slippery and on standing on them the player will slide towards an end. Second, there exists a secret passage that provides an alternate route to the goal which is for the user to find out. This passage is initially invisible but the user can find it by search around all across the platform. It will automatically show itself when the user comes in contact with it.